+++ src/main/java/nextstep/courses/domain/Width.java
@@ -1,14 +1,13 @@
 package nextstep.courses.domain;
 
 public class Width {
-    private final Integer value;
+    private final Double value;
 
-    public Width(Integer width) {
+    public Width(Double width) {
         if (width < 300) {
             throw new IllegalArgumentException("너비는 300px 이상이어야 합니다.");
         }
 
         this.value = width;
-
     }
 }
